Use raw loader for .html templates
On build I was getting an $compile:tpload error. So, I changed templateURL to template
and used raw loader (which came pre-installed with angular-webpack) to load the
templates.

One thing to note it that there was no error after the webpack build, only after
I deployed it to Firebase did I see the error.